requirements.txt
tweepy
beautifulsoup4
requests
schedule
python-dotenv
pillow
loguru
pytz
